move ReceiveEvent/SendEvent from Client namespace to Socket namespace per review
