+++ generators/typescript/sdk/client-class-generator/src/GeneratedWebsocketSocketClassImpl.ts
@@ -72,6 +72,7 @@ export class GeneratedWebsocketSocketClassImpl implements GeneratedWebsocketSock
             statements: [
                 this.generateSocketArgs(context),
                 this.generateSocketResponse(context),
+                ...this.generateEventTypeAliases(context),
                 this.generateSocketHandlers(context)
             ]
         };
@@ -181,6 +182,29 @@ export class GeneratedWebsocketSocketClassImpl implements GeneratedWebsocketSock
             isExported: true,
             type: getTextOfTsNode(this.getUnionedNodeForOrigin(context, "server"))
         };
+    }
+
+    private generateEventTypeAliases(context: SdkContext): TypeAliasDeclarationStructure[] {
+        const aliases: TypeAliasDeclarationStructure[] = [];
+        const serverMessages = this.getMessagesForOrigin("server");
+        if (serverMessages.length > 0) {
+            aliases.push({
+                kind: StructureKind.TypeAlias,
+                name: "ReceiveEvent",
+                isExported: true,
+                type: getTextOfTsNode(this.getUnionedNodeForOrigin(context, "server"))
+            });
+        }
+        const clientMessages = this.getMessagesForOrigin("client");
+        if (clientMessages.length > 0) {
+            aliases.push({
+                kind: StructureKind.TypeAlias,
+                name: "SendEvent",
+                isExported: true,
+                type: getTextOfTsNode(this.getUnionedNodeForOrigin(context, "client"))
+            });
+        }
+        return aliases;
     }
 
     private generateSocketHandlers(context: SdkContext): TypeAliasDeclarationStructure {
