+++ seed/ts-sdk/websocket-bearer-auth/websockets/src/api/resources/realtime/client/Socket.ts
@@ -14,6 +14,15 @@ export declare namespace RealtimeSocket {
         | SeedWebsocketBearerAuth.ReceiveSnakeCase
         | SeedWebsocketBearerAuth.ReceiveEvent2
         | SeedWebsocketBearerAuth.ReceiveEvent3;
+    export type ReceiveEvent =
+        | SeedWebsocketBearerAuth.ReceiveEvent
+        | SeedWebsocketBearerAuth.ReceiveSnakeCase
+        | SeedWebsocketBearerAuth.ReceiveEvent2
+        | SeedWebsocketBearerAuth.ReceiveEvent3;
+    export type SendEvent =
+        | SeedWebsocketBearerAuth.SendEvent
+        | SeedWebsocketBearerAuth.SendSnakeCase
+        | SeedWebsocketBearerAuth.SendEvent2;
     type EventHandlers = {
         open?: () => void;
         message?: (message: Response) => void;
