+++ seed/ts-sdk/websocket/no-serde/src/api/resources/realtime/client/Socket.ts
@@ -15,6 +15,13 @@ export declare namespace RealtimeSocket {
         | SeedWebsocket.ReceiveEvent2
         | SeedWebsocket.ReceiveEvent3
         | SeedWebsocket.ErrorEvent;
+    export type ReceiveEvent =
+        | SeedWebsocket.ReceiveEvent
+        | SeedWebsocket.ReceiveSnakeCase
+        | SeedWebsocket.ReceiveEvent2
+        | SeedWebsocket.ReceiveEvent3
+        | SeedWebsocket.ErrorEvent;
+    export type SendEvent = SeedWebsocket.SendEvent | SeedWebsocket.SendSnakeCase | SeedWebsocket.SendEvent2;
     type EventHandlers = {
         open?: () => void;
         message?: (message: Response) => void;
