+++ seed/ts-sdk/websocket/serde/src/api/resources/realtime/client/Socket.ts
@@ -19,6 +19,13 @@ export declare namespace RealtimeSocket {
         | SeedWebsocket.ReceiveEvent2
         | SeedWebsocket.ReceiveEvent3
         | SeedWebsocket.ErrorEvent;
+    export type ReceiveEvent =
+        | SeedWebsocket.ReceiveEvent
+        | SeedWebsocket.ReceiveSnakeCase
+        | SeedWebsocket.ReceiveEvent2
+        | SeedWebsocket.ReceiveEvent3
+        | SeedWebsocket.ErrorEvent;
+    export type SendEvent = SeedWebsocket.SendEvent | SeedWebsocket.SendSnakeCase | SeedWebsocket.SendEvent2;
     type EventHandlers = {
         open?: () => void;
         message?: (message: Response) => void;
